% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyAlloc.R
\name{Trt_Prob}
\alias{Trt_Prob}
\title{Creates a vector of probabilities of being assigned to each treatment group}
\usage{
Trt_Prob(imbalances, alpha)
}
\arguments{
\item{imbalances}{a vector}

\item{alpha}{a numeric number}
}
\description{
Creates a vector of probabilities of being assigned to each treatment group
}
