% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyAlloc.R
\name{WeiImb}
\alias{WeiImb}
\title{Creates a weighted imbalance score for each treatment arm
*NOTE: the default for site here is FALSE, but one can specify site=TRUE in the argument
to account for site imbalance. Note that the weight for site has to be placed in the
third element of the 'weight' vector (and can be omitted when site=FALSE)*}
\usage{
WeiImb(df, N, covars, weight, newsub, site = FALSE)
}
\arguments{
\item{df}{data frame.The profiles of existing subjects in the study.}

\item{N}{numeric variable. The number of treatment arms in the study.}

\item{covars}{a vector. All the names of covariates in the study.}

\item{weight}{a vector. The weights for overall study, with-in stratum, (site),
and factors/covariates, in this particular order}

\item{newsub}{a vector. The factor profile of the new subject entering the study.}

\item{site}{a boolean value. Whether or not to account for site imbalance (default:FALSE)}
}
\value{
A vector of weighted imbalance scores for each treatment arm is produced (a total of N scores)
}
\description{
Creates a weighted imbalance score for each treatment arm
*NOTE: the default for site here is FALSE, but one can specify site=TRUE in the argument
to account for site imbalance. Note that the weight for site has to be placed in the
third element of the 'weight' vector (and can be omitted when site=FALSE)*
}
