% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyAlloc.R
\name{DyAlloc}
\alias{DyAlloc}
\title{provides the final treatment group assignment}
\usage{
DyAlloc(df, N, covars, weight, newsub, site = FALSE)
}
\arguments{
\item{df}{data frame}

\item{N}{numeric variable}

\item{covars}{a vector}

\item{weight}{a vector}

\item{newsub}{a vector}

\item{site}{a boolean value}
}
\description{
provides the final treatment group assignment
}
