% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyAlloc.R
\name{MarImb}
\alias{MarImb}
\title{Creates a marginal imbalance score for each treatment arm of N treatment groups}
\usage{
MarImb(df, N)
}
\arguments{
\item{df}{data frame}

\item{N}{numeric variable}
}
\description{
Creates a marginal imbalance score for each treatment arm of N treatment groups
}
