% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyAlloc.R
\name{Trt_Prob}
\alias{Trt_Prob}
\title{Creates a vector of probabilities of being assigned to each treatment group.
This function uses the imbalance scores created from the function WeiImb as input.
*NOTE: if the imbalance scores are the same, then they have equal probability of
being assigned to. E.g. if we get two equal least weighted imbalance scores, they each
have a probability of 0.5 for a subject to be assigned to.*}
\usage{
Trt_Prob(imbalances, alpha)
}
\arguments{
\item{imbalances}{a vector. Imbalance scores across the N treatment groups
(generally, the scores we get from WeiImb would be plugged in here)}

\item{alpha}{a numeric number. The "second best probability" parameter, generally alpha is set to 0.2.
This means that the treat group a subject assign to gets the probability of 0.8.}
}
\value{
A vector of probabilities of being assigned to each treatment group.
}
\description{
Creates a vector of probabilities of being assigned to each treatment group.
This function uses the imbalance scores created from the function WeiImb as input.
*NOTE: if the imbalance scores are the same, then they have equal probability of
being assigned to. E.g. if we get two equal least weighted imbalance scores, they each
have a probability of 0.5 for a subject to be assigned to.*
}
