% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DyAlloc.R
\name{WeiImb}
\alias{WeiImb}
\title{Creates a weighted imbalance score for each treatment arm}
\usage{
WeiImb(df, N, covars, weight, newsub, site = FALSE)
}
\arguments{
\item{df}{data frame}

\item{N}{numeric variable}

\item{covars}{a vector}

\item{weight}{a vector}

\item{newsub}{a vector}

\item{site}{a boolean value}
}
\description{
Creates a weighted imbalance score for each treatment arm
}
